requirements.txt
Django
djangorestframework
django-crispy-forms
django-filter
django-request
django-guardian
Markdown
Pigments
coreapi
psycopg2
PyPDF2
textract
matplotlib
jsonschema
nltk
numpy
pandas
pdfminer.six==20181108
spacy==2.3.5
pyresparser
gunicorn
